+++ hc/accounts/views.py
@@ -355,6 +355,11 @@ def unsubscribe_reports(request, username):
     except signing.BadSignature:
         return render(request, "bad_link.html")
 
+    # Some email servers open links in emails to check for malicious content.
+    # To work around this, we serve a form that auto-submits with JS.
+    if "ask" in request.GET and request.method != "POST":
+        return render(request, "accounts/unsubscribe_submit.html")
+
     user = User.objects.get(username=username)
     profile = Profile.objects.for_user(user)
     profile.reports_allowed = False